# src/apidoc.json
{
  "swagger": "2.0",
  "info": {
    "title": "Local Marketplace",
    "description": "This is set of API's for business to integrate our services in their applications.",
    "contact": {
      "email": "abcd@example.com"
    },
    "license": {
      "name": "BSD License"
    },
    "version": "v1.0.0"
  },
  "host": "localhost:3001",
  "schemes": [
    "http"
  ],
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Delivery": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "Seller": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Delivery": []
    },
    {
      "Seller": []
    }
  ],
  "paths": {
...
    "/seller/{id}/": {
      "get": {
        "operationId": "seller_read",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ReadOnlySellerSerializer"
            }
          }
        },
        "tags": [
          "seller"
        ]
      },
      "put": {
        "operationId": "seller_update",
        "description": "",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WriteOnlySellerSerializer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ReadOnlySellerSerializer"
            }
          }
        },
        "tags": [
          "seller"
        ]
      },
      "patch": {
        "operationId": "seller_partial_update",
        "description": "",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WriteOnlySellerSerializer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ReadOnlySellerSerializer"
            }
          }
        },
        "tags": [
          "seller"
        ]
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "A UUID string identifying this seller.",
          "required": true,
          "type": "string"
        }
      ]
    }
  },
  "definitions": {
...
    "ReadOnlySellerSerializer": {
      "required": [
        "address",
        "email",
        "phone",
        "pan",
        "gst"
      ],
      "type": "object",
      "properties": {
        "id": {
          "title": "Id",
          "type": "string"
        },
        "firstName": {
          "title": "Firstname",
          "type": "string",
          "maxLength": 150
        },
        "address": {
          "title": "Address",
          "type": "object",
          "properties": {
            "pincode": {
              "title": "Pin Code",
              "type": "string"
            },
            "state": {
              "title": "State",
              "type": "string"
            },
            "city": {
              "title": "City",
              "type": "string"
            },
            "landmark": {
              "title": "Landmark",
              "type": "string"
            },
            "houseNo": {
              "title": "House No",
              "type": "string"
            },
            "streetName": {
              "title": "Street Name",
              "type": "string"
            }
          },
          "required": [
            "pincode",
            "state",
            "city"
          ]
        },
        "email": {
          "title": "Email",
          "type": "string",
          "format": "email",
          "maxLength": 254,
          "minLength": 1
        },
        "phone": {
          "title": "Phone",
          "type": "string",
          "maxLength": 13,
          "minLength": 1
        },
        "pan": {
          "title": "Pan",
          "type": "string",
          "maxLength": 13,
          "minLength": 1
        },
        "gst": {
          "title": "Gst",
          "type": "string",
          "maxLength": 13,
          "minLength": 1
        }
      }
    },
    "WriteOnlySellerSerializer": {
      "required": [
        "address",
        "email",
        "phone",
        "pan",
        "gst"
      ],
      "type": "object",
      "properties": {
        "id": {
          "title": "Id",
          "type": "string"
        },
        "firstName": {
          "title": "Firstname",
          "type": "string",
          "maxLength": 150
        },
        "address": {
          "title": "Address",
          "type": "object",
          "properties": {
            "pincode": {
              "title": "Pin Code",
              "type": "string"
            },
            "state": {
              "title": "State",
              "type": "string"
            },
            "city": {
              "title": "City",
              "type": "string"
            },
            "landmark": {
              "title": "Landmark",
              "type": "string"
            },
            "houseNo": {
              "title": "House No",
              "type": "string"
            },
            "streetName": {
              "title": "Street Name",
              "type": "string"
            }
          },
          "required": [
            "pincode",
            "state",
            "city"
          ]
        },
        "email": {
          "title": "Email",
          "type": "string",
          "format": "email",
          "maxLength": 254,
          "minLength": 1
        },
        "phone": {
          "title": "Phone",
          "type": "string",
          "maxLength": 13,
          "minLength": 1
        },
        "pan": {
          "title": "Pan",
          "type": "string",
          "maxLength": 13,
          "minLength": 1
        },
        "gst": {
          "title": "Gst",
          "type": "string",
          "maxLength": 13,
          "minLength": 1
        }
      }
    }
  }
}